here is the code for task 1 of version2 and there is still something wrong file syscall.d contains some dirty data

+++ kernel/defs.h
@@ -82,7 +82,7 @@ void            panic(char*) __attribute__((noreturn));
 void            printfinit(void);
 
 // proc.c
-int             getprocs(void);
+uint64          getprocs(void);
 int             cpuid(void);
 void            exit(int);
 int             fork(void);
